[refactor] CommentController Interface 분리로 Swagger 설정 분리
