+++ src/main/java/com/bbteam/budgetbuddies/domain/comment/controller/CommentController.java
@@ -1,29 +1,21 @@
 package com.bbteam.budgetbuddies.domain.comment.controller;
 
+
 import com.bbteam.budgetbuddies.domain.comment.dto.CommentRequestDto;
 import com.bbteam.budgetbuddies.domain.comment.dto.CommentResponseDto;
-import com.bbteam.budgetbuddies.domain.comment.service.CommentService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-public class CommentController {
-
-    private final CommentService commentService;
-
-    // user, discountInfo 인증 어노테이션 추후 추가 예정
-
+public interface CommentController {
     @Operation(summary = "[User] 특정 할인 정보 게시글에 댓글달기", description = "특정 할인 정보 게시글에 댓글을 다는 API입니다.")
     @ApiResponses({
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON200", description = "OK, 성공"),
@@ -33,13 +25,10 @@ public class CommentController {
             @Parameter(name = "discountInfoId", description = "댓글을 다는 할인 정보 게시글 id입니다. requestBody"),
             @Parameter(name = "content", description = "댓글 내용입니다. requestBody"),
     })
-    @PostMapping("/discounts/comments/add")
-    public ResponseEntity<CommentResponseDto.DiscountInfoSuccessDto> saveDiscountInfoComment(
-            @RequestParam("userId") Long userId,
-            @RequestBody CommentRequestDto.DiscountInfoCommentDto discountInfoCommentDto){
-        CommentResponseDto.DiscountInfoSuccessDto dto = commentService.saveDiscountComment(userId, discountInfoCommentDto);
-        return ResponseEntity.ok(dto);
-    }
+    ResponseEntity<CommentResponseDto.DiscountInfoSuccessDto> saveDiscountInfoComment(
+            Long userId,
+            CommentRequestDto.DiscountInfoCommentDto discountInfoCommentDto);
+
 
     @Operation(summary = "[User] 특정 할인 정보 게시글의 댓글 조회하기", description = "특정 할인 정보 게시글의 댓글을 가져오는 API입니다.")
     @ApiResponses({
@@ -50,13 +39,9 @@ public class CommentController {
             @Parameter(name = "page", description = "페이징을 위한 페이지 번호입니다. 0부터 시작합니다. parameter"),
             @Parameter(name = "size", description = "페이징을 위한 페이지 사이즈입니다. default는 20입니다. parameter")
     })
-    @GetMapping("/discounts/comments/get")
-    public ResponseEntity<Page<CommentResponseDto.DiscountInfoCommentDto>> findAllByDiscountInfo(
-            @RequestParam("discountInfoId") Long discountInfoId,
-            @PageableDefault(size = 20, page = 0) Pageable pageable){
-        Page<CommentResponseDto.DiscountInfoCommentDto> result = commentService.findByDiscountInfoWithPaging(discountInfoId, pageable);
-        return ResponseEntity.ok(result);
-    }
+    ResponseEntity<Page<CommentResponseDto.DiscountInfoCommentDto>> findAllByDiscountInfo(
+            Long discountInfoId,
+            Pageable pageable);
 
     @Operation(summary = "[User] 특정 지원 정보 게시글에 댓글달기", description = "특정 지원 정보 게시글에 댓글을 다는 API입니다.")
     @ApiResponses({
@@ -67,13 +52,9 @@ public class CommentController {
             @Parameter(name = "supportInfoId", description = "댓글을 다는 지원 정보 게시글 id입니다. requestBody"),
             @Parameter(name = "content", description = "댓글 내용입니다. requestBody"),
     })
-    @PostMapping("/supports/comments/add")
-    public ResponseEntity<CommentResponseDto.SupportInfoSuccessDto> saveSupportInfoComment(
-            @RequestParam("userId") Long userId,
-            @RequestBody CommentRequestDto.SupportInfoCommentDto supportInfoCommentDto){
-        CommentResponseDto.SupportInfoSuccessDto dto = commentService.saveSupportComment(userId, supportInfoCommentDto);
-        return ResponseEntity.ok(dto);
-    }
+    ResponseEntity<CommentResponseDto.SupportInfoSuccessDto> saveSupportInfoComment(
+            Long userId,
+            CommentRequestDto.SupportInfoCommentDto supportInfoCommentDto);
 
     @Operation(summary = "[User] 특정 지원 정보 게시글의 댓글 조회하기", description = "특정 지원 정보 게시글의 댓글을 가져오는 API입니다.")
     @ApiResponses({
@@ -86,13 +67,9 @@ public class CommentController {
 
 
     })
-    @GetMapping("/supports/comments/get")
-    public ResponseEntity<Page<CommentResponseDto.SupportInfoCommentDto>> findAllBySupportInfo(
-            @RequestParam("supportInfoId") Long supportInfoId,
-            @PageableDefault(size = 20, page = 0)Pageable pageable){
-        Page<CommentResponseDto.SupportInfoCommentDto> result = commentService.findBySupportInfoWithPaging(supportInfoId, pageable);
-        return ResponseEntity.ok(result);
-    }
+    ResponseEntity<Page<CommentResponseDto.SupportInfoCommentDto>> findAllBySupportInfo(
+            Long supportInfoId,
+            Pageable pageable);
 
     @Operation(summary = "[User] 특정 댓글 삭제하기", description = "특정 댓글을 삭제하는 API입니다.")
     @ApiResponses({
@@ -100,16 +77,7 @@ public class CommentController {
     })
     @Parameters({
             @Parameter(name = "commentId", description = "삭제할 댓글 id 입니다. parameter")
-
-
     })
     @GetMapping("/comments/delete")
-    public ResponseEntity<String> deleteComment(@RequestParam("commentId") Long commentId) {
-        commentService.deleteComment(commentId);
-        return ResponseEntity.ok("ok");
-    }
-
-
-
-
+    ResponseEntity<String> deleteComment(Long commentId);
 }
