+++ src/main/java/com/bbteam/budgetbuddies/domain/comment/controller/CommentControllerImpl.java
@@ -1,0 +1,64 @@
+package com.bbteam.budgetbuddies.domain.comment.controller;
+
+import com.bbteam.budgetbuddies.domain.comment.dto.CommentRequestDto;
+import com.bbteam.budgetbuddies.domain.comment.dto.CommentResponseDto;
+import com.bbteam.budgetbuddies.domain.comment.service.CommentService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.Parameters;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+public class CommentControllerImpl implements CommentController{
+
+    private final CommentService commentService;
+
+    @PostMapping("/discounts/comments/add")
+    public ResponseEntity<CommentResponseDto.DiscountInfoSuccessDto> saveDiscountInfoComment(
+            @RequestParam("userId") Long userId,
+            @RequestBody CommentRequestDto.DiscountInfoCommentDto discountInfoCommentDto){
+        CommentResponseDto.DiscountInfoSuccessDto dto = commentService.saveDiscountComment(userId, discountInfoCommentDto);
+        return ResponseEntity.ok(dto);
+    }
+
+
+    @GetMapping("/discounts/comments/get")
+    public ResponseEntity<Page<CommentResponseDto.DiscountInfoCommentDto>> findAllByDiscountInfo(
+            @RequestParam("discountInfoId") Long discountInfoId,
+            @PageableDefault(size = 20, page = 0) Pageable pageable){
+        Page<CommentResponseDto.DiscountInfoCommentDto> result = commentService.findByDiscountInfoWithPaging(discountInfoId, pageable);
+        return ResponseEntity.ok(result);
+    }
+
+
+    @PostMapping("/supports/comments/add")
+    public ResponseEntity<CommentResponseDto.SupportInfoSuccessDto> saveSupportInfoComment(
+            @RequestParam("userId") Long userId,
+            @RequestBody CommentRequestDto.SupportInfoCommentDto supportInfoCommentDto){
+        CommentResponseDto.SupportInfoSuccessDto dto = commentService.saveSupportComment(userId, supportInfoCommentDto);
+        return ResponseEntity.ok(dto);
+    }
+
+
+    @GetMapping("/supports/comments/get")
+    public ResponseEntity<Page<CommentResponseDto.SupportInfoCommentDto>> findAllBySupportInfo(
+            @RequestParam("supportInfoId") Long supportInfoId,
+            @PageableDefault(size = 20, page = 0)Pageable pageable){
+        Page<CommentResponseDto.SupportInfoCommentDto> result = commentService.findBySupportInfoWithPaging(supportInfoId, pageable);
+        return ResponseEntity.ok(result);
+    }
+
+
+    public ResponseEntity<String> deleteComment(@RequestParam("commentId") Long commentId) {
+        commentService.deleteComment(commentId);
+        return ResponseEntity.ok("ok");
+    }
+
+}
